Switch to Dgraph
