+++ requirements.txt
@@ -1,12 +1,6 @@
-amun
-aiogremlin==3.2.6rc1
-boto3
-cython
-goblin==2.1.0
-gremlinpython==3.2.6
-requests
-thoth-common
-ujson
-uvloop
 voluptuous
-yarl<1.2
+boto3
+thoth-common
+amun
+pydgraph
+
