+++ thoth/storages/graph/__init__.py
@@ -1,17 +1,21 @@
-"""Graph database models and adapter."""
+#!/usr/bin/env python3
+# thoth-storages
+# Copyright(C) 2019 Fridolin Pokorny
+#
+# This program is free software: you can redistribute it and / or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-from .janusgraph import GraphDatabase
-from .models import CreatesStack
-from .models import DependsOn
-from .models import DependsOn
-from .models import HasVersion
-from .models import IsPartOf
-from .models import Solved
-from .models import Package
-from .models import PythonPackageVersion
-from .models import RPMPackageVersion
-from .models import RuntimeEnvironment
-from .models import Observed
-from .models import HardwareInformation
-from .models import SoftwareStackObservation
-from .models import ALL_MODELS
+"""A graph database adapter for communicating with dgraph via gRPC."""
+
+
+from .dgraph import GraphDatabase
